Rename files from *config* to *setting*. Rename 'getAll' functions to convention
