+++ api/serverless.yml
@@ -17,7 +17,7 @@ provider:
       Resource: "arn:aws:dynamodb:${opt:region, self:provider.region}:*"
 
 functions:
-  venues_list:
+  get_venues_list:
     handler: venues/list.list
     events:
       - http:
@@ -49,7 +49,7 @@ functions:
           method: delete
           cors: true
 
-  content_list:
+  get_content_list:
     handler: content/list.list
     events:
       - http:
@@ -81,24 +81,24 @@ functions:
           method: delete
           cors: true
 
-  config_list:
-    handler: config/list.list
+  get_settings_list:
+    handler: setting/list.list
     events:
       - http:
           path: api/configs
           method: get
           cors: true
 
-  create_config:
-    handler: config/create.create
+  create_setting:
+    handler: setting/create.create
     events:
       - http:
           path: api/configs
           method: post
           cors: true
 
-  update_config:
-    handler: config/update.update
+  update_setting:
+    handler: setting/update.update
     events:
       - http:
           path: api/configs/{id}
