+++ api/setting/list.js
@@ -5,6 +5,6 @@ const responseHelper = require('../helpers/http_response_helper');
 
 module.exports.list = (event, context, callback) => {
     dbHelper.findAll(process.env.CONFIGS_TABLE)
-        .then(config => callback(null, responseHelper.createSuccessfulResponse(config)))
+        .then(setting => callback(null, responseHelper.createSuccessfulResponse(setting)))
         .fail(error => callback(null, responseHelper.createResponseWithError(500, error.message)));
 };
