+++ api/setting/update.js
@@ -1,6 +1,6 @@
 'use strict';
 
-let Config = require('./../entities/config');
+let Config = require('./../entities/setting');
 
 const dynamodb = require('../dynamodb');
 const responseHelper = require('../helpers/http_response_helper');
