+++ api/test/get_settings_list.js
@@ -3,8 +3,8 @@
 require('./helpers/test_provider_configurator').configure();
 const DatabaseCleaner = require('./helpers/database_cleaner');
 
-const getAllFunction = require('../config/list.js');
-const createFunction = require('../config/create.js');
+const getAllFunction = require('../setting/list.js');
+const createFunction = require('../setting/create.js');
 const mochaPlugin = require('serverless-mocha-plugin');
 
 const expect = mochaPlugin.chai.expect;
@@ -16,7 +16,7 @@ const MultiOperationHelper = require('./helpers/multi_operation_test_helper')
 
 const idLength = 36;
 
-describe('config_list', () => {
+describe('get_settings_list', () => {
     before((done) => {
         DatabaseCleaner.cleanDatabase().then(() => done());
     });
@@ -33,7 +33,7 @@ describe('config_list', () => {
         });
     });
 
-    it('Should display list with 1 config', () => {
+    it('Should display list with 1 setting', () => {
         let config = {name: 'New year'};
 
         let expectations = (body, response) => {
